requirements.txt
cachetools==3.1.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.2
future==0.17.1
google-api-core==1.14.2
google-auth==1.6.3
google-cloud-vision==0.39.0
googleapis-common-protos==1.6.0
grpcio==1.23.0
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
line-bot-sdk==1.8.0
MarkupSafe==1.1.1
protobuf==3.9.1
pyasn1==0.4.7
pyasn1-modules==0.2.6
python-dotenv==0.10.1
pytz==2019.2
redis==3.3.8
requests==2.21.0
rsa==4.0
six==1.12.0
urllib3==1.24.2
Werkzeug==0.15.5
